bfd: Appease a CHERI warning by adding an intermediate cast
It is ok for this pointer to be untagged as the size_t is just being
passed to a pointer-hashing function to get a hash value.

+++ bfd/dwarf2.c
@@ -1140,7 +1140,7 @@ static hashval_t
 hash_abbrev (const void *p)
 {
   const struct abbrev_offset_entry *ent = p;
-  return htab_hash_pointer ((void *) ent->offset);
+  return htab_hash_pointer ((void *)(uintptr_t) ent->offset);
 }
 
 static int
